requirements.txt
requests
python-dotenv
notion-client
Flask
PyPDF2
openai
google-api-python-client==2.122.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
